payment setup

+++ database/migrations/2025_07_22_104949_add_task_id_to_files_table.php
@@ -12,7 +12,9 @@ return new class extends Migration
     public function up(): void
     {
         Schema::table('files', function (Blueprint $table) {
-            $table->foreignId('task_id')->nullable()->constrained()->onDelete('cascade');
+            if (!Schema::hasColumn('files', 'task_id')) {
+                $table->foreignId('task_id')->nullable()->constrained()->onDelete('cascade');
+            }
         });
     }
